Geometry Nodes: Implemented an expression node that evaluates a text string as a mathematical expression.
Doing any sort of calculation is quite painful with simple primative nodes, often requiring multiple nodes for even a simple calculation. The expression node simplifies graph construction reduces node clutter and makes the intent of calculations clearer. As a bonus, expression strings can be passed into node groups allowing generic groups for things like graph drawing or creating volumes of rotation.

The expression node takes a string as input that is similar to a mathematical experssion, and very similar to the way expression are written in computer code. For instance "x * sin(theta) + y", where x, y and theta are inputs to the node. Float, int, vector and boolean inputs are accepted, and the result can be one of the same type. Inputs to the node can be renamed to make the expression clearer. A smart naming feature gets the name of an input socket from the socket connecting to it.

Mathematical operators +, -, * (multiply), /, ^ (power) and % (modulo) are supported. A wide range of triganometric, rounding, logical and vector functions are available. Constants pi, tau and e can be used. Individual components of vectors can be accessed, and vectors created from components (e.g. vector(a.x, b.y, a.z) where a and b are vector inputs.

Example expressions

x
x + 2 * (y + z)
2 * asin(to_radians(theta))
x^2 + y^2
x < 10 and x > -10
if(x > 4, y, z)
2 * cross(v1, v2) + v3

The node works by tokenizing the expression via a recursive descent parser, then re-organizing the tokens into a postfix form which can be evaluated efficiantly with a simple stack based machine. The re-organization phase also checks the operand types and selects the correct token for the operation.